Crate sui_crypto

Modules

bls12381bls12381

ed25519ed25519

multisiged25519 or secp256r1 or secp256k1 or zklogin

passkeypasskey

secp256k1secp256k1

secp256r1secp256r1

simpleed25519 or secp256r1 or secp256k1 or zklogin

zkloginzklogin

Structs

SignatureError

Signature errors.

UserSignatureVerifiered25519 or secp256r1 or secp256k1 or zklogin

Verifier that will verify all UserSignature variants

Traits

Signer

Sign the provided message bytestring using Self (e.g. a cryptographic key or connection to an HSM), returning a digital signature.

SuiSigner

Interface for signing user transactions and messages in Sui

SuiVerifier

Interface for verifying user transactions and messages in Sui

Verifier

Verify the provided message bytestring using Self (e.g. a public key)